+++ ast/src/list.h
@@ -7,8 +7,8 @@
 #include <stdalign.h> /* alignas */
 
 struct kn_list {
-	alignas(8) unsigned refcount;
-	unsigned length;
+	alignas(8) size_t refcount;
+	size_t length;
 	kn_value *elements;
 };
 
@@ -39,13 +39,13 @@ void kn_list_dump(const struct kn_list *list, FILE *out);
 
 struct kn_string *kn_list_join(const struct kn_list *list, const struct kn_string *sep);
 struct kn_list *kn_list_concat(struct kn_list *lhs, struct kn_list *rhs);
-struct kn_list *kn_list_repeat(struct kn_list *list, unsigned amount);
+struct kn_list *kn_list_repeat(struct kn_list *list, size_t amount);
 bool kn_list_equal(const struct kn_list *lhs, const struct kn_list *rhs);
-struct kn_list *kn_list_get(struct kn_list *list, unsigned start, unsigned length);
+struct kn_list *kn_list_get(struct kn_list *list, size_t start, size_t length);
 struct kn_list *kn_list_set(
 	struct kn_list *list,
-	unsigned start,
-	unsigned length,
+	size_t start,
+	size_t length,
 	struct kn_list *replacement
 );
